requirements.txt
appdirs==1.4.3
arrow==0.10.0
click==6.7
Flask>=0.12.3
gevent==1.2.1
greenlet==0.4.12
iso8601==0.1.11
itsdangerous==0.24
Jinja2>=2.10.1
leancloud-sdk==1.10.0
MarkupSafe==1.0
packaging==16.8
pyparsing==2.2.0
python-dateutil==2.6.0
qiniu==7.1.2
requests>=2.20.0
six==1.10.0
Werkzeug>=0.14
nose==1.3.7
Tornado==5.1.1
flask_cors